refactor context menus to extend popovers
