+++ editor/src/ui/popoverMenus/ContextMenu.js
@@ -1,6 +1,6 @@
 import {ContextMenuItem} from "./ContextMenuItem.js";
 import {ContextMenuSubmenuItem} from "./ContextMenuSubmenuItem.js";
-import {Button} from "../Button.js";
+import {Popover} from "./Popover.js";
 
 /**
  * @typedef {object} ContextMenuOptions
@@ -14,20 +14,6 @@ import {Button} from "../Button.js";
  * @typedef {object} ContextMenuItemClickEvent
  * @property {ContextMenuItem} item
  * @property {function() : void} preventMenuClose
- */
-
-/**
- * `"clamp"`: Clamp the menu to the screen.
- * `"flip"`: Moves the menu so that the corner is now on the opposite side.
- * If an element is provided, rather than a position, the menu will also be moved
- * in a way to keep the element visible.
- * @typedef {"flip" | "clamp"} ContextMenuSetPosClampMode
- */
-
-/**
- * @typedef {"left" | "center" | "right"} ContextMenuSetPosHorizontalCorner
- * @typedef {"top" | "center" | "bottom"} ContextMenuSetPosVerticalCorner
- * @typedef {ContextMenuSetPosHorizontalCorner | ContextMenuSetPosVerticalCorner | `${ContextMenuSetPosHorizontalCorner} ${ContextMenuSetPosVerticalCorner}` | `${ContextMenuSetPosVerticalCorner} ${ContextMenuSetPosHorizontalCorner}`} ContextMenuSetPosCorner
  */
 
 /**
@@ -46,7 +32,7 @@ import {Button} from "../Button.js";
  * @property {ContextMenuStructure | (function(): Promise<ContextMenuStructure>) | function(): ContextMenuStructure} [submenu=null] The submenu structure to show on hover.
  */
 
-export class ContextMenu {
+export class ContextMenu extends Popover {
 	/**
 	 * @param {import("./PopoverManager.js").PopoverManager} manager
 	 * @param {ContextMenuOptions} opts
@@ -55,18 +41,15 @@ export class ContextMenu {
 		parentMenu = null,
 		structure = null,
 	} = {}) {
-		this.manager = manager;
+		super(manager);
 		this.parentMenu = parentMenu;
-		this.el = document.createElement("div");
-		this.el.classList.add("contextMenu");
-		document.body.appendChild(this.el);
 
 		/** @type {Array<ContextMenuItem>} */
 		this.addedItems = [];
 		this.activeSubmenuItem = null;
 		/** @type {ContextMenu?} */
 		this.currentSubmenu = null;
-		/** @type {ContextMenuSetPosOpts?} */
+		/** @type {import("./Popover.js").ContextMenuSetPosOpts?} */
 		this.lastPosArguments = null;
 
 		this.hasResevedIconSpaceItem = false;
@@ -77,14 +60,12 @@ export class ContextMenu {
 	}
 
 	destructor() {
+		super.destructor();
 		this.removeSubmenu();
 		for (const item of this.addedItems) {
 			item.destructor();
 		}
 		this.addedItems = [];
-		if (this.el) {
-			if (this.el.parentElement) this.el.parentElement.removeChild(this.el);
-		}
 	}
 
 	removeSubmenu() {
@@ -94,108 +75,13 @@ export class ContextMenu {
 		}
 	}
 
-	/**
-	 * @typedef {object} ContextMenuSetPosOpts
-	 * @property {number} [x]
-	 * @property {number} [y]
-	 * @property {HTMLElement | Button | ContextMenuItem} [item]
-	 * @property {ContextMenuSetPosCorner} [corner]
-	 * @property {ContextMenuSetPosClampMode} [clampMode]
-	 * @property {boolean} [preventElementCover] When true, will move the menu out of the way
-	 * when the menu would otherwise cover the element as a result of the clamp mode. Only works
-	 * with clamp mode "flip".
+		/**
+	 * @param  {import("./Popover.js").ContextMenuSetPosOpts} options
 	 */
-
-	/**
-	 * @param  {ContextMenuSetPosOpts} options
-	 */
-	setPos(options) {
-		this.lastPosArguments = {...options};
-
-		let x = options.x;
-		let y = options.y;
-		let corner = options.corner;
-		let clampMode = options.clampMode;
-		let preventElementCover = options.preventElementCover;
-		let el = null;
-
-		if (options.item instanceof ContextMenuItem) {
-			el = options.item.el;
-			if (!clampMode) clampMode = "flip";
-			if (preventElementCover == undefined) preventElementCover = true;
+		setPos(options) {
+			this.lastPosArguments = {...options};
+			super.setPos(options);
 		}
-		if (options.item instanceof Button) {
-			el = options.item.el;
-			if (!corner) corner = "top left";
-		}
-		let elRect = null;
-		if (el instanceof HTMLElement) {
-			elRect = el.getBoundingClientRect();
-			if (!corner) corner = "center";
-			const corenerArgs = corner.split(" ");
-			const castCornerArgs = /** @type {(ContextMenuSetPosHorizontalCorner | ContextMenuSetPosVerticalCorner)[]} */ (corenerArgs);
-			/** @type {ContextMenuSetPosHorizontalCorner} */
-			let horizontalCorner = "center";
-			/** @type {ContextMenuSetPosVerticalCorner} */
-			let verticalCorner = "center";
-			if (castCornerArgs.includes("left")) horizontalCorner = "left";
-			if (castCornerArgs.includes("right")) horizontalCorner = "right";
-			if (castCornerArgs.includes("top")) verticalCorner = "top";
-			if (castCornerArgs.includes("bottom")) verticalCorner = "bottom";
-
-			if (horizontalCorner == "center") {
-				x = elRect.x + elRect.width / 2;
-			} else if (horizontalCorner == "left") {
-				x = elRect.x;
-			} else if (horizontalCorner == "right") {
-				x = elRect.right;
-			}
-			if (verticalCorner == "center") {
-				y = elRect.y + elRect.height / 2;
-			} else if (verticalCorner == "top") {
-				y = elRect.top;
-			} else if (verticalCorner == "bottom") {
-				y = elRect.bottom;
-			}
-
-			if (!clampMode) clampMode = "clamp";
-		}
-
-		if (x == undefined) x = 0;
-		if (y == undefined) y = 0;
-
-		if (!clampMode) clampMode = "flip";
-
-		const bounds = this.el.getBoundingClientRect();
-		if (clampMode == "flip") {
-			if (x + bounds.width > window.innerWidth) {
-				x -= bounds.width;
-				if (preventElementCover && elRect) {
-					x -= elRect.width;
-				}
-			}
-			if (y + bounds.height > window.innerHeight) {
-				y -= bounds.height;
-				if (preventElementCover && elRect) {
-					y -= elRect.height;
-				}
-			}
-		} else if (clampMode == "clamp") {
-			const deltaX = x + bounds.width - window.innerWidth;
-			if (deltaX > 0) {
-				x -= deltaX;
-				x = Math.max(0, x);
-			}
-			const deltaY = y + bounds.height - window.innerHeight;
-			if (deltaY > 0) {
-				y -= deltaY;
-				y = Math.max(0, y);
-			}
-		}
-
-		this.el.style.left = x + "px";
-		this.el.style.top = y + "px";
-	}
 
 	/**
 	 * @param {ContextMenuStructure} structure
@@ -270,11 +156,6 @@ export class ContextMenu {
 		} else {
 			this.close();
 		}
-	}
-
-	close() {
-		this.manager.onPopoverClosed(this);
-		this.destructor();
 	}
 
 	updateHasReservedIconSpaceItem() {
